chore: upgrade deps and speedup on linux (#4)
* centralize metadata builders and add fetch/parse tracing

* tune fetch client and filter listing

* gate fast-alloc to linux wheels

* use manylinux 2_28 for linux wheels

+++ src/iter/peek.rs
@@ -9,7 +9,7 @@ use chrono::{DateTime, Utc};
 use nexrad_data::volume::{Header, Record};
 use nexrad_decode::messages::MessageContents;
 use object_store::path::Path as ObjectPath;
-use object_store::{GetOptions, GetRange, ObjectStore};
+use object_store::{GetOptions, GetRange, ObjectStore, ObjectStoreExt};
 use pyo3::prelude::*;
 use pyo3::types::PyDateTime;
 use std::sync::Arc;
@@ -448,7 +448,7 @@ async fn fetch_range(
     let _permit = FETCH_SEMAPHORE.acquire().await.expect("semaphore closed");
 
     let options = GetOptions {
-        range: Some(GetRange::Bounded(start..end)),
+        range: Some(GetRange::Bounded(start as u64..end as u64)),
         ..Default::default()
     };
